[PrayerTimesKit] Restrict acos input to the domain to prevent NaN
Returning NaN causes problems later in processing, eventually leading
to a crash. There are different opinions on when prayer times should
be for latitudes that observe these behaviors- this is not one of them.

I decided to check this in because beyond crashing, the app won't
open again until the days shorten to be able to observe a typical
solar cycle.