[PrayerTimesWKExtension] Start RollingPrayersView from active prayer
The original reason for this code (being removed in this commit)
was for the time between midnight and Fajr, show Fajr as the first
prayer, as opposed to Isha from the day before. Showing Isha for
the day before results in the following format:

    Isha
    ## Today
    Fajr
    ...

which I originally thought might be confusing, but I no longer
think that's the case. Additionally, due to TimelineView
behaviors, the code behaved unexpectedly:

In "normal" display mode, the context date was that of Isha,
therefore "today" that is calculated is actually yesterday,
and we observe the same behavior as with this commit.

In "Always-On" display mode, the context date is
approximately the current date, which means that between
midnight and Fajr, there is no active prayer shown.

In other words, this commit changes the "Always-On" behavior
to match that of "normal" display mode.